+++ python/ray/rllib/agents/pg/pg_policy.py
@@ -5,7 +5,7 @@ from __future__ import print_function
 import ray
 from ray.rllib.evaluation.postprocessing import compute_advantages, \
     Postprocessing
-from ray.rllib.evaluation.tf_policy_template import build_tf_policy
+from ray.rllib.policy.tf_policy_template import build_tf_policy
 from ray.rllib.evaluation.sample_batch import SampleBatch
 from ray.rllib.utils import try_import_tf
