+++ python/ray/rllib/evaluation/__init__.py
@@ -1,10 +1,10 @@
 from ray.rllib.evaluation.episode import MultiAgentEpisode
-from ray.rllib.evaluation.policy_evaluator import PolicyEvaluator
+from ray.rllib.policy.policy_evaluator import PolicyEvaluator
 from ray.rllib.evaluation.interface import EvaluatorInterface
-from ray.rllib.evaluation.policy import Policy
-from ray.rllib.evaluation.policy_graph import PolicyGraph
-from ray.rllib.evaluation.tf_policy import TFPolicy
-from ray.rllib.evaluation.tf_policy_graph import TFPolicyGraph
+from ray.rllib.policy.policy import Policy
+from ray.rllib.policy.policy_graph import PolicyGraph
+from ray.rllib.policy.tf_policy import TFPolicy
+from ray.rllib.policy.tf_policy_graph import TFPolicyGraph
 from ray.rllib.evaluation.torch_policy import TorchPolicy
 from ray.rllib.evaluation.torch_policy_graph import TorchPolicyGraph
 from ray.rllib.evaluation.sample_batch import SampleBatch, MultiAgentBatch
