+++ python/ray/rllib/policy/dynamic_tf_policy.py
@@ -6,9 +6,9 @@ from collections import OrderedDict
 import logging
 import numpy as np
 
-from ray.rllib.evaluation.policy import Policy
+from ray.rllib.policy.policy import Policy
 from ray.rllib.evaluation.sample_batch import SampleBatch
-from ray.rllib.evaluation.tf_policy import TFPolicy
+from ray.rllib.policy.tf_policy import TFPolicy
 from ray.rllib.models.catalog import ModelCatalog
 from ray.rllib.utils.annotations import override
 from ray.rllib.utils import try_import_tf
